requirements.txt
gensim==4.3.2
firebase-admin==6.5.0
nltk~=3.8.1
google-cloud-core==2.4.1
google-auth==2.29.0
google-cloud-firestore==2.15.0
google-api-python-client==2.124.0
scipy==1.12.0